Likes

+++ src/main/java/com/example/ourstoryapp/da/LikesRepository.java
@@ -5,10 +5,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 import com.example.ourstoryapp.domain.Likes;
 
-
 @RepositoryRestResource
-public interface LikesRepository extends JpaRepository <Likes,Long> {
-
-	
+public interface LikesRepository extends JpaRepository<Likes, Integer> {
 
 }
